charts refactor
