+++ web/netlify/functions/market-chart.mts
@@ -1,3 +1,4 @@
+import { Market } from "@/hooks/useMarket";
 import { SupportedChain } from "@/lib/chains";
 import { createClient } from "@supabase/supabase-js";
 import { Address } from "viem";
@@ -6,14 +7,45 @@ import { getChartData } from "./utils/getChartData";
 
 const supabase = createClient(process.env.VITE_SUPABASE_PROJECT_URL!, process.env.VITE_SUPABASE_API_KEY!);
 
-export function getMarketChartKeyValueHash(
-  marketId: Address | "%",
-  chainId: SupportedChain,
-  dayCount: number,
-  intervalSeconds: number,
-  endDateParam: number | undefined,
-) {
-  return `market_chart_${marketId}_${chainId}_${dayCount}_${intervalSeconds}_${endDateParam || "latest"}`;
+export function getMarketChartKeyValueHash(marketId: Address | "%", chainId: SupportedChain | "%") {
+  return `market_chart_hour_data_${marketId}_${chainId}`;
+}
+
+async function fetchChartData(market: Market) {
+  const hashKey = getMarketChartKeyValueHash(market.id, market.chainId);
+
+  const { data: cachedData, error: cacheError } = await supabase
+    .from("key_value")
+    .select("value")
+    .eq("key", hashKey)
+    .single();
+
+  if (
+    cacheError?.code === "PGRST116" ||
+    (cachedData?.value && Date.now() - cachedData.value.timestamp > 5 * 60 * 1000)
+  ) {
+    // data not found or older than 5 minutes
+    const chartData = await getChartData(market);
+    const cacheData = { chartData, timestamp: Date.now(), marketId: market.id };
+
+    // store and return
+    const { error: upsertError } = await supabase.from("key_value").upsert(
+      {
+        key: hashKey,
+        value: cacheData,
+      },
+      { onConflict: "key" },
+    );
+
+    if (upsertError) {
+      console.error("Cache upsert failed:", upsertError);
+    }
+
+    return chartData;
+  }
+
+  // return cached data
+  return cachedData?.value.chartData;
 }
 
 export default async (req: Request) => {
@@ -32,11 +64,6 @@ export default async (req: Request) => {
 
   const marketId = params.get("marketId") as Address;
   const chainId = Number(params.get("chainId")) as SupportedChain;
-  const dayCount = Number.parseInt(params.get("dayCount") || "0", 10);
-  const intervalSeconds = Number.parseInt(params.get("intervalSeconds") || "0", 10);
-  const endDateParam = Number(params.get("endDate")) * 1000;
-
-  const endDate = endDateParam ? new Date(endDateParam) : undefined;
 
   if (!marketId || !chainId) {
     return new Response(JSON.stringify({ error: "Missing required parameters" }), {
@@ -60,46 +87,7 @@ export default async (req: Request) => {
       });
     }
 
-    // check if we have data on cache
-    const hashKey = getMarketChartKeyValueHash(marketId, chainId, dayCount, intervalSeconds, endDateParam);
-
-    const { data: cachedData, error: cacheError } = await supabase
-      .from("key_value")
-      .select("value")
-      .eq("key", hashKey)
-      .single();
-
-    if (
-      cacheError?.code === "PGRST116" ||
-      (cachedData?.value && Date.now() - cachedData.value.timestamp > 5 * 60 * 1000)
-    ) {
-      // data not found or older than 5 minutes
-      const chartData = await getChartData(market, dayCount, intervalSeconds, endDate);
-      const cacheData = { chartData, timestamp: Date.now(), marketId: market.id };
-
-      // store and return
-      const { error: upsertError } = await supabase.from("key_value").upsert(
-        {
-          key: hashKey,
-          value: cacheData,
-        },
-        { onConflict: "key" },
-      );
-
-      if (upsertError) {
-        console.error("Cache upsert failed:", upsertError);
-      }
-
-      return new Response(JSON.stringify(chartData), {
-        status: 200,
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-    }
-
-    // return cached data
-    return new Response(JSON.stringify(cachedData?.value.chartData), {
+    return new Response(JSON.stringify(await fetchChartData(market)), {
       status: 200,
       headers: {
         "Content-Type": "application/json",
