+++ web/src/components/Market/MarketChart.tsx
@@ -1,5 +1,5 @@
 import { ChartData } from "@/hooks/chart/useChartData";
-import { fetchChartData, useChartData } from "@/hooks/chart/useChartData";
+import { fetchFullChartData, useChartData } from "@/hooks/chart/useChartData";
 import { useIsSmallScreen } from "@/hooks/useIsSmallScreen";
 import { Market } from "@/hooks/useMarket";
 import { ExportIcon, QuestionIcon } from "@/lib/icons";
@@ -260,7 +260,7 @@ function MarketChart({ market }: { market: Market }) {
 
   const exportData = async () => {
     // Use resolved date for export if available
-    const { chartData, timestamps }: ChartData = await fetchChartData(market, 365 * 10, 60 * 60 * 24, undefined);
+    const { chartData, timestamps }: ChartData = await fetchFullChartData(market);
     const series = getSeries(market, chartData);
     const headers = [
       {
